Rename reader files for semantic clarity
- Renamed parquet_reader.rs to parquet_query_engine.rs (high-level query logic)
- Renamed shared_parquet_reader.rs to parquet_io_layer.rs (low-level I/O operations)
- Renamed unified_sstable_reader.rs to sst_query_engine.rs (high-level SST queries)
- Renamed shared_sst_reader.rs to sst_io_layer.rs (low-level SST I/O)

Added comprehensive documentation explaining the layered architecture:
- Query engines handle business logic, search, filtering, scoring
- I/O layers handle caching, memory mapping, cloud storage optimization

This clarifies that these are not duplicates but complementary layers.

+++ src/core/bloom/factory.rs
@@ -91,7 +91,7 @@ impl BloomFilterStrategy for NoOpBloomFilter {
         0
     }
     
-    fn hash_count(&self) -> u32 {
+    fn hash_count(&self) -> usize {
         0
     }
