+++ src/core/bloom/strategies/simple.rs
@@ -61,8 +61,8 @@ impl BloomFilterStrategy for SimpleBloomFilter {
         self.bits.len()
     }
     
-    fn hash_count(&self) -> u32 {
-        self.num_hashes
+    fn hash_count(&self) -> usize {
+        self.num_hashes as usize
     }
     
     fn serialize(&self) -> Result<Vec<u8>> {
